+++ models/config.go
@@ -47,22 +47,23 @@ type Cameras struct {
 }
 
 type Alerts struct {
-	General    General      `fig:"general" json:"general,omitempty" doc:"Common alert settings"`
-	Quiet      Quiet        `fig:"quiet" json:"quiet,omitempty" doc:"Alert quiet periods"`
-	Zones      Zones        `fig:"zones" json:"zones,omitempty" doc:"Allow/Block zones from alerting"`
-	Labels     Labels       `fig:"labels" json:"labels,omitempty" doc:"Allow/Block labels from alerting"`
-	SubLabels  Labels       `fig:"sublabels" json:"sublabels,omitempty" doc:"Allow/Block sublabels from alerting"`
-	AppriseAPI []AppriseAPI `fig:"apprise-api" json:"apprise-api,omitempty" doc:"Apprise API notification settings"`
-	Discord    []Discord    `fig:"discord" json:"discord,omitempty" doc:"Discord notification settings"`
-	Gotify     []Gotify     `fig:"gotify" json:"gotify,omitempty" doc:"Gotify notification settings"`
-	Matrix     []Matrix     `fig:"matrix" json:"matrix,omitempty" doc:"Matrix notification settings"`
-	Mattermost []Mattermost `fig:"mattermost" json:"mattermost,omitempty" doc:"Mattermost notification settings"`
-	Ntfy       []Ntfy       `fig:"ntfy" json:"ntfy,omitempty" doc:"Ntfy notification settings"`
-	Pushover   []Pushover   `fig:"pushover" json:"pushover,omitempty" doc:"Pushover notification settings"`
-	Signal     []Signal     `fig:"signal" json:"signal,omitempty" doc:"Signal notification settings"`
-	SMTP       []SMTP       `fig:"smtp" json:"smtp,omitempty" doc:"SMTP notification settings"`
-	Telegram   []Telegram   `fig:"telegram" json:"telegram,omitempty" doc:"Telegram notification settings"`
-	Webhook    []Webhook    `fig:"webhook" json:"webhook,omitempty" doc:"Webhook notification settings"`
+	General      General      `fig:"general" json:"general,omitempty" doc:"Common alert settings"`
+	Quiet        Quiet        `fig:"quiet" json:"quiet,omitempty" doc:"Alert quiet periods"`
+	Zones        Zones        `fig:"zones" json:"zones,omitempty" doc:"Allow/Block zones from alerting"`
+	Labels       Labels       `fig:"labels" json:"labels,omitempty" doc:"Allow/Block labels from alerting"`
+	SubLabels    Labels       `fig:"sublabels" json:"sublabels,omitempty" doc:"Allow/Block sublabels from alerting"`
+	LicensePlate LicensePlate `fig:"license_plate" json:"license_plate,omitempty" doc:"License plate recognition settings"`
+	AppriseAPI   []AppriseAPI `fig:"apprise-api" json:"apprise-api,omitempty" doc:"Apprise API notification settings"`
+	Discord      []Discord    `fig:"discord" json:"discord,omitempty" doc:"Discord notification settings"`
+	Gotify       []Gotify     `fig:"gotify" json:"gotify,omitempty" doc:"Gotify notification settings"`
+	Matrix       []Matrix     `fig:"matrix" json:"matrix,omitempty" doc:"Matrix notification settings"`
+	Mattermost   []Mattermost `fig:"mattermost" json:"mattermost,omitempty" doc:"Mattermost notification settings"`
+	Ntfy         []Ntfy       `fig:"ntfy" json:"ntfy,omitempty" doc:"Ntfy notification settings"`
+	Pushover     []Pushover   `fig:"pushover" json:"pushover,omitempty" doc:"Pushover notification settings"`
+	Signal       []Signal     `fig:"signal" json:"signal,omitempty" doc:"Signal notification settings"`
+	SMTP         []SMTP       `fig:"smtp" json:"smtp,omitempty" doc:"SMTP notification settings"`
+	Telegram     []Telegram   `fig:"telegram" json:"telegram,omitempty" doc:"Telegram notification settings"`
+	Webhook      []Webhook    `fig:"webhook" json:"webhook,omitempty" doc:"Webhook notification settings"`
 }
 
 type General struct {
@@ -75,8 +76,14 @@ type General struct {
 	MaxSnapRetry     int    `fig:"max_snap_retry,omitempty" json:"max_snap_retry" doc:"Maximum number of retry attempts when snapshot is not ready yet" default:"10"`
 	NotifyOnce       bool   `fig:"notify_once,omitempty"  json:"notify_once" enum:"true,false" doc:"Only notify once per event (For app mode: events)" default:false`
 	NotifyDetections bool   `fig:"notify_detections,omitempty" json:"notify_detections" enum:"true,false" doc:"Enable notifications on detection (For app mode: reviews)" default:false`
-	RecheckDelay     int    `fig:"recheck_delay" json:"recheck_delay" default:"0" doc:"Delay before re-checking event details from Frigate"`
-	AudioOnly        string `fig:"audio_only" json:"audio_only" enum:"allow,drop" doc:"Allow/Drop events that only contain audio detections" default:"allow"`
+	RecheckDelay     int    `fig:"recheck_delay" json:"recheck_delay,omitempty" default:"0" doc:"Delay before re-checking event details from Frigate"`
+	AudioOnly        string `fig:"audio_only" json:"audio_only,omitempty" enum:"allow,drop" doc:"Allow/Drop events that only contain audio detections" default:"allow"`
+}
+
+type LicensePlate struct {
+	Enabled bool     `fig:"enabled" json:"enabled,omitempty" enum:"true,false" doc:"Enable waiting for license plate recognition when car & license plate are detected" default:false`
+	Allow   []string `fig:"allow" json:"allow,omitempty" doc:"List of license plates to allow alerts from" default:[]`
+	Block   []string `fig:"block" json:"block,omitempty" doc:"List of license plates to always block" default:[]`
 }
 
 type Quiet struct {